- Summary: New feature of the data binding UI
- Bugzilla Bug (s) Resolved: 148235

- Description:
new feature:
1.Create a Table with existing Dataset.
2.Dragging and dropping data from palette will open the dialog where user can create e new data binding.
3.Double Click on Data Item in Editor,Show the dialog with existing values and changes in the values is made possible.
4.Right Click Menu Option on Data Item in editor: Change Data Column.
5.Property Editor Binding Tab behavior changed.

148325:Adding a binding column with the same expression as a previously added bindind column doesn't work

- Tests Description: Manual test.

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Added:
@addlist

- Files Edited:
@changelist

- Files Deleted:
@deletelist

+++ UI/org.eclipse.birt.report.designer.ui/src/org/eclipse/birt/report/designer/internal/ui/dialogs/TableOptionDialog.java
@@ -294,8 +294,8 @@ public class TableOptionDialog extends BaseDialog
 		if ( insertTable )
 		{
 			setResult( new Object[]{
-					rowCount,
-					columnCount,
+					(Integer)rowCount,
+					(Integer)columnCount,
 					dataSetCombo.getItem( dataSetCombo.getSelectionIndex( ) )
 							.toString( )
 			} );
